requirements.txt
torch==1.3.0
numpy==1.16.2
scipy==1.2.1
open3d-python==0.7.0.0
scikit-image==0.15.0
